requirements.txt
yfinance>=0.2.36
pandas>=2.0
numpy>=1.26
ta>=0.11.0
requests>=2.31
feedparser>=6.0
python-dotenv>=1.0
